Project 5/requirements.txt
pandas==2.1.3
numpy==1.24.3
matplotlib==3.8.2
scikit-learn==1.3.2
torch==2.1.0
ccxt==4.0.91
python-dotenv==1.0.0 